Fix: patch dts and power_debug for kernel build success

+++ drivers/soc/sprd/power_debug/sprd_pdbg_sharkl5pro.c
@@ -10,6 +10,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
+#include <linux/types.h>
 #include <linux/module.h>
 #include <linux/device.h>
 #include <linux/of_platform.h>
@@ -556,7 +557,7 @@ static int sprd_pdbg_sharkl5pro_probe(struct platform_device *pdev)
 	pcfg = sprd_pdbg_parse_cfg(&pdev->dev,
 		sharkl5pro_pdbg_desc.ap_intc_num);
 	if (IS_ERR(pcfg))
-		return (int)pcfg;
+		return (uintptr_t)pcfg;
 
 	pdbg_sharkl5pro.pdbg = sprd_power_debug_register(&pdev->dev,
 				&sharkl5pro_pdbg_desc,	pcfg);
